feat: refactor metrics, add `log_metadata` arg, allow more metric args
