+++ src/euroeval/metrics/__init__.py
@@ -1,0 +1,6 @@
+"""All the metrics used in EuroEval."""
+
+from .huggingface import *  # noqa: F403
+from .llm_as_a_judge import *  # noqa: F403
+from .pipeline import *  # noqa: F403
+from .speed import *  # noqa: F403
